Safari-611.1.10/LayoutTests/editing/pasteboard/paste-cocoa-writer-markup-with-system-fonts-expected.txt

This tests pasting a markup from Cocoa HTML Writer with system font references. WebKit should strip system font names away upon paste.
To manually test, click on "start" first then trigger paste from menu or callout bar or press cmd+v

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS event.clipboardData.getData('text/html') is markup
PASS editor.innerHTML.includes('.AppleSystemUIFont') is true
PASS editor.innerHTML.includes('.SFUI-Regular') is true
PASS successfullyParsed is true

TEST COMPLETE
Start
